pyproject.toml
[tool.poetry]
name = "data-agent"
version = "0.1.0"
description = "Census data fetching and database management"
authors = ["User <user@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
